+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Status (in loggers list) have now their own column.
 - The `README.md` displays badges.
 ### Removed
+- DecaLog admin pages no longer render emoji.
 - The section "install from GitHub" has been removed from `README.md`.
 ### Fixed
 - There was a wrong url for the EFF website in the `readme.txt`.